requirements.txt
aiohttp>=3.9.0
alembic==1.17.0
annotated-types==0.7.0
anthropic==0.71.0
anyio==4.11.0
asyncpg==0.30.0
attrs==25.4.0
Authlib==1.6.5
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
coverage==7.11.0
cryptography==46.0.2
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.119.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.1
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
iniconfig==2.1.0
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
Mako==1.3.10
MarkupSafe==3.0.3
mcp==1.17.0
numpy==2.3.3
openai==2.4.0
packaging==25.0
pluggy==1.6.0
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.0
pydantic-settings==2.11.0
pydantic_core==2.41.1
Pygments==2.19.2
pyparsing==3.2.5
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
pytest-html==4.1.1
pytest-json-report==1.5.0
pytest-metadata==3.1.1
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
sniffio==1.3.1
SQLAlchemy==2.0.44
sse-starlette==3.0.2
starlette==0.48.0
tavily-python==0.5.0
tenacity>=8.2.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.37.0
uvloop==0.21.0
validators==0.35.0
watchfiles==1.1.0
weaviate-client==4.17.0
websockets==15.0.1
apscheduler>=3.10.0
pypdf==5.1.0
plaid-python>=23.0.0
